Introduce Spacemacs-based Emacs config as my primary
Spacemacs is, at the moment, my favorite Emacs distribution, and where
most of my time has been spent. Thus it is my most custom.

My current config is only broken down in Org at a high level, in the
interest of time. Even so, it's easier to understand than in the form
Spacemacs generates it as.