Introduce YASnippet for GTD project templating
It's easy to forget parts of the GTD natural planning model, and I'd
like to avoid fishing out my copy of GTD when project planning. This
commit uses YASnippet, the default template system for Spacemacs, for
a single snippet specific to new GTD projects, with useful formatting
and links to the backing concepts in place.

The new snippet points to a file outside this repository. I consider
this a problem but one of low importance.